refactor(syscall): split network syscalls
